% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_list.R
\name{bullet_list}
\alias{bullet_list}
\title{Bullet List}
\usage{
bullet_list(items, bullet = "\\t-", show = TRUE, html = FALSE)
}
\arguments{
\item{items}{TBD}

\item{bullet}{TBD}

\item{show}{TBD}

\item{html}{TBD}
}
\description{
Form a properly formatted bullet list from a list of items
}
\examples{
TBD
}
