% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{CreateDocument}
\alias{CreateDocument}
\title{Document}
\usage{
CreateDocument(paragraphs, show = TRUE, html = TRUE, timestamp = TRUE)
}
\arguments{
\item{show}{TBD}

\item{html}{TBD}

\item{timestamp}{TBD}

\item{sentences}{TBD}
}
\description{
Form a properly formatted document from a list of paragraphs, images, and bullet-lists
}
\examples{
TBD

}
