% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{CreateHeader}
\alias{CreateHeader}
\title{Header}
\usage{
CreateHeader(text, h = 1, show = TRUE, html = TRUE)
}
\arguments{
\item{show}{TBD}

\item{html}{TBD}

\item{items}{TBD}

\item{bullet}{TBD}
}
\description{
Create a header
}
\examples{
TBD
}
