% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{CreateImageTag}
\alias{CreateImageTag}
\title{Image}
\usage{
CreateImageTag(url, alt = NULL, show = TRUE)
}
\arguments{
\item{alt}{TBD}

\item{src}{TBD}
}
\description{
Format an HTML image tag
}
\examples{
TBD
}
