% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{CreateParagraph}
\alias{CreateParagraph}
\title{Paragraph}
\usage{
CreateParagraph(sentences)
}
\arguments{
\item{sentences}{TBD}
}
\description{
Form a properly formatted paragraph from a list of sentences
}
\examples{
TBD

}
