% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{CreateSentence}
\alias{CreateSentence}
\title{Sentence}
\usage{
CreateSentence(fragments = list("this is the first",
  "and this is the second"), punct = ".")
}
\arguments{
\item{fragments}{TBD}

\item{punct}{TBD}
}
\description{
Form a properly formatted sentence from a list of fragments
}
\examples{
CreateSentence(list("one more question:", "how do you feel"), punct = '?')
}
