% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{CreateTimestamp}
\alias{CreateTimestamp}
\title{Timestamp}
\usage{
CreateTimestamp(html = TRUE)
}
\arguments{
\item{html}{TBD}
}
\description{
A statement indicating the script completion time
}
\examples{
TBD

}
