% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_list.R, R/header.R
\name{CreateUnorderedList}
\alias{CreateUnorderedList}
\alias{CreateUnorderedList}
\title{Bullet List}
\usage{
CreateUnorderedList(text, h = 1, show = TRUE, html = TRUE)

CreateUnorderedList(text, h = 1, show = TRUE, html = TRUE)
}
\arguments{
\item{show}{TBD}

\item{html}{TBD}

\item{items}{TBD}

\item{bullet}{TBD}

\item{items}{TBD}

\item{bullet}{TBD}

\item{show}{TBD}

\item{html}{TBD}
}
\description{
Form a properly formatted bullet list from a list of items

Create a header
}
\examples{
TBD
TBD
}
