% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_url.R
\name{CreateURLTag}
\alias{CreateURLTag}
\title{Create a URL (href) tag (Deprecated)}
\usage{
CreateURLTag(href, text = "Link")
}
\arguments{
\item{href}{A character value for the URL (href) attribute of the desired link}

\item{text}{The text value for the link}
}
\description{
Format an url tag
}
\examples{
CreateURLTag("www.google.com", "Let me google that for you...")
}
