% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasis.R
\name{emphasis}
\alias{emphasis}
\title{Add emphasis}
\usage{
emphasis(text, type = "bold")
}
\arguments{
\item{text}{The text string to be emphasized}

\item{type}{The type of emphasis ("bold" or "italic")}
}
\description{
Bold or italicize text to add emphasis to a string
}
