% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metadata.R
\name{eval_meta}
\alias{eval_meta}
\title{Parse and evaluate metadata within an Rmarkdown document}
\usage{
eval_meta(meta)
}
\arguments{
\item{meta}{the metadata field value as return}
}
\value{
Returns properly evauated metadata for the current Rmarkdown document
}
\description{
\code{eval_meta()} creates properly formatted HTML and metadata from an R markdown file. See \href{https://stackoverflow.com/questions/45646135/evaluating-r-code-in-yaml-header}{this thread} for why this is necessary.
}
\examples{
eval_meta(rmarkdown::metadata$date)
}
