% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_conditional_refac.R
\name{fragment_conditional}
\alias{fragment_conditional}
\title{Create a fragment based on a boolean condition}
\usage{
fragment_conditional(items, conditions, fallback = NA)
}
\arguments{
\item{items}{An unnamed list of character values in the desired order.}

\item{conditions}{A list of boolean values corresponding to each value in items. When more than one value is true, the first is returned.}

\item{fallback}{The character string to return when the list is empty or no conditions are met.}
}
\description{
Return a single character value from a pre-defined set based on a corresponding set of conditions
}
\examples{
x <- 10
example_items <- list("an increase", "a decrease", "no change")
example_bool <- list(x > 0, x < 0, x == 0)
fragment_conditional(example_items, example_bool)
}
