% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_series.R
\name{fragment_series}
\alias{fragment_series}
\title{Fragment: Simple Series}
\usage{
fragment_series(items, conj = "and", fallback = "nothing")
}
\arguments{
\item{items}{An unnamed list of character values in the desired order}

\item{conj}{The conjunction to use: "and", "or", &c.}

\item{fallback}{The character string to return when the list is empty.}
}
\description{
Format a list of character values as a simple series
}
\examples{
example_items <- list("shoes", "ships", "sealing wax")
fragment_series(example_items)
}
