% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_tag.R
\name{header_tag}
\alias{header_tag}
\title{Add a header tag}
\usage{
header_tag(text, h = 1)
}
\arguments{
\item{text}{The text value for the header}

\item{h}{The header level}
}
\description{
Create a header tag (H1 - H6)
}
