% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_tag.R
\name{link_tag}
\alias{link_tag}
\title{Add a link tag}
\usage{
link_tag(text, url)
}
\arguments{
\item{text}{The text of the link}

\item{url}{The URL the link points to}
}
\description{
Format an HTML image tag
}
