% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ordered_list.R
\name{ordered_list}
\alias{ordered_list}
\title{Create an ordered list}
\usage{
ordered_list(items)
}
\arguments{
\item{items}{A list or vector of items (in desired order)}
}
\description{
Generate markdown for a properly formatted numbered list from a vector of items
}
\examples{
ordered_list(items = list("DC Brau", "Helllbender", "Three Star", "Right Proper", "Blue Jacket", "Public Option"))
}
