% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R, R/paragraph.R
\name{paragraph}
\alias{paragraph}
\alias{paragraph}
\title{Document}
\usage{
paragraph(sentences, show = TRUE)

paragraph(sentences, show = TRUE)
}
\arguments{
\item{sentences}{TBD}

\item{show}{TBD}

\item{html}{TBD}

\item{timestamp}{TBD}

\item{sentences}{TBD}

\item{show}{TBD}
}
\description{
Form a properly formatted document from a list of paragraphs and bullet-lists

Form a properly formatted paragraph from a list of sentences
}
\examples{
TBD

TBD

}
