% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{sentence}
\alias{sentence}
\title{Sentence}
\usage{
sentence(fragments = list("this is the first", "and this is the second"),
  show = TRUE, punct = ".")
}
\arguments{
\item{fragments}{TBD}

\item{show}{TBD}

\item{punct}{TBD}
}
\description{
Form a properly formatted sentence from a list of fragments
}
\examples{
TBD
}
