% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unordered_list.R
\name{unordered_list}
\alias{unordered_list}
\title{Create an unordered list}
\usage{
unordered_list(items)
}
\arguments{
\item{items}{A list or vector of items (in desired order)}
}
\description{
Generate markdown for a properly formatted bullet list from a vector of items
}
\examples{
unordered_list(items = list("DC Brau", "Helllbender", "Three Star", "Right Proper", "Blue Jacket", "Public Option"))
}
