% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_count.R
\name{CreateCountFragment}
\alias{CreateCountFragment}
\title{Fragment: Count of values in a vector}
\usage{
CreateCountFragment(x, units, fallback = NA, unique = FALSE)
}
\arguments{
\item{x}{A vector of any class.}

\item{units}{A vector of length 2 indicating the singular and plural of x's units (e.g., c("country","countries")).}

\item{fallback}{The character string to return in case of error}

\item{unique}{Should only unique values be counted?}
}
\description{
Return a statment summarising the number of items in a set
}
\examples{
CreateCountFragment(c('A','A','B','B','C','C','C'), units = c('letter', 'letters'), unique = TRUE)
}
