% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{CreateDocument}
\alias{CreateDocument}
\title{Document}
\usage{
CreateDocument(paragraphs)
}
\arguments{
\item{paragraphs}{A list of paragraph objects}
}
\description{
Form a properly formatted document from a list of paragraphs, images, and bullet-lists
}
