% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{CreateHeader}
\alias{CreateHeader}
\title{Add a header tag (Deprecated)}
\usage{
CreateHeader(text, h = 1)
}
\arguments{
\item{text}{The text value for the header}

\item{h}{The header level}
}
\description{
Create a header tag (H1 - H6)
}
