% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{CreateParagraph}
\alias{CreateParagraph}
\title{Compile sentences into a paragraph}
\usage{
CreateParagraph(sentences)
}
\arguments{
\item{sentences}{A list of sentences to be compiled into a paragraph.}
}
\description{
Form a properly formatted paragraph from a list of sentences
}
