% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{CreateSentence}
\alias{CreateSentence}
\title{Compile fragments into a sentence (Deprecated)}
\usage{
CreateSentence(fragments = list("this is the first",
  "and this is the second"), punct = ".")
}
\arguments{
\item{fragments}{A list of fragments to be compiled into a sentence.}

\item{punct}{The punctuation used at the end of the sentence.}
}
\description{
Form a properly formatted sentence from a list of fragments
}
\examples{
CreateSentence(list("one more question:", "how do you feel"), punct = '?')
}
