% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_list.R
\name{CreateUnorderedList}
\alias{CreateUnorderedList}
\title{Bullet List}
\usage{
CreateUnorderedList(items, show = TRUE, html = TRUE)
}
\arguments{
\item{items}{TBD}

\item{show}{TBD}

\item{html}{TBD}

\item{bullet}{TBD}
}
\description{
Form a properly formatted bullet list from a list of items
}
\examples{
TBD
}
