% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_list.R
\name{CreateUnorderedList}
\alias{CreateUnorderedList}
\title{Create an unordered list}
\usage{
CreateUnorderedList(items)
}
\arguments{
\item{items}{A list or vector of items (in desired order)}
}
\description{
Form a properly formatted bullet list from a vector of items
}
\examples{
CreateUnorderedList(items = list("DC Brau", "Helllbender", "Three Star", "Right Proper", "Blue Jacket", "Public Option"))
}
