% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_url.R
\name{CreateURLTag}
\alias{CreateURLTag}
\title{URL Fragment}
\usage{
CreateURLTag(href, text = "Link", show = TRUE)
}
\arguments{
\item{href}{TBD}

\item{text}{TBD}
}
\description{
Format an url tag
}
\examples{
TBD
}
