% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_count.R
\name{fragment_count}
\alias{fragment_count}
\title{Fragment: Count of values in a vector}
\usage{
fragment_count(x, units, fallback = NA, unique = FALSE)
}
\arguments{
\item{x}{A vector of any class.}

\item{units}{A vector of length 2 indicating the singular and plural of x's units (e.g., c("country","countries")).}

\item{fallback}{The character string to return in case of error}

\item{unique}{Should only unique values be counted?}
}
\description{
Return a statment summarising the number of items in a set
}
\examples{
x <- 10
example_items <- list("an increase", "a decrease", "no change")
example_bool <- list(x > 0, x < 0, x == 0)
fragment_conditional(example_items, example_bool)
}
